maj accounts + users

+++ components/Profil.jsx
@@ -13,9 +13,15 @@ export default function Profil({ user }){
     const [rgpd, setRgpd] = useState(false)
     
     const getUserInfos = async () => {
-        const geter = await fetch(`/api/users/${user.id}`)
-        const json = await geter.json()
-        setUserData(json[0])
+        if(user.type == "Administrateur" || user.type == "DR"){
+            const geter = await fetch(`/api/accounts/${user.id}`)
+            const json = await geter.json()
+            setUserData(json[0])
+        } else {
+            const geter = await fetch(`/api/users/${user.id}`)
+            const json = await geter.json()
+            setUserData(json[0])
+        }
     }
 
     const handleChange = async (event) => {
@@ -42,22 +48,42 @@ export default function Profil({ user }){
         const { motDePasse, motDePasse2 } = passw
         if(motDePasse && motDePasse2){
             if(motDePasse.length > 5){
-                const fetcher = await fetch('/api/users/lib/update-password', {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify({
-                        motDePasse: motDePasse,
-                        userId: user.id
+                if(user.type == "Administrateur" || user.type == "DR"){
+                    // const fetcher = await fetch('/api/accounts/lib/update-password', {
+                    //     method: 'POST',
+                    //     headers: {
+                    //         'Content-Type': 'application/json'
+                    //     },
+                    //     body: JSON.stringify({
+                    //         motDePasse: motDePasse,
+                    //         userId: user.id
+                    //     })
+                    // })
+                    // const json = await fetcher.json();
+                    // if(json.id){
+                    //     setNotif({
+                    //         text: 'Le mot de passe a bien été modififé !',
+                    //         icon: 'done'
+                    //     })                      
+                    // }
+                } else {
+                    const fetcher = await fetch('/api/users/lib/update-password', {
+                        method: 'POST',
+                        headers: {
+                            'Content-Type': 'application/json'
+                        },
+                        body: JSON.stringify({
+                            motDePasse: motDePasse,
+                            userId: user.id
+                        })
                     })
-                })
-                const json = await fetcher.json();
-                if(json.id){
-                    setNotif({
-                        text: 'Le mot de passe a bien été modififé !',
-                        icon: 'done'
-                    })                      
+                    const json = await fetcher.json();
+                    if(json.id){
+                        setNotif({
+                            text: 'Le mot de passe a bien été modififé !',
+                            icon: 'done'
+                        })                      
+                    }
                 }
             }
             else{
@@ -78,21 +104,39 @@ export default function Profil({ user }){
     const deleteAccount = async () => {
         if(rgpd){
             try {
-                const response = await fetch(`/api/users/delete/?id=${user.id}`, {
-                    method: 'DELETE',
-                    headers: {
-                        'Content-Type': 'application/json',
-                    },
-                });
-        
-                if (!response.ok) {
-                    throw new Error(`Erreur HTTP: ${response.status}`);
+                if(user.type == "Administrateur" || user.type == "DR"){
+                    // const response = await fetch(`/api/accounts/delete/?id=${user.id}`, {
+                    //     method: 'DELETE',
+                    //     headers: {
+                    //         'Content-Type': 'application/json',
+                    //     },
+                    // });
+            
+                    // if (!response.ok) {
+                    //     throw new Error(`Erreur HTTP: ${response.status}`);
+                    // }
+            
+                    // const result = await response.json();
+                    // const unlog = await fetch('/api/logout')
+                    
+                    // window.location.href = "/"
+                } else {
+                    const response = await fetch(`/api/users/delete/?id=${user.id}`, {
+                        method: 'DELETE',
+                        headers: {
+                            'Content-Type': 'application/json',
+                        },
+                    });
+            
+                    if (!response.ok) {
+                        throw new Error(`Erreur HTTP: ${response.status}`);
+                    }
+            
+                    const result = await response.json();
+                    const unlog = await fetch('/api/logout')
+                    
+                    window.location.href = "/"
                 }
-        
-                const result = await response.json();
-                const unlog = await fetch('/api/logout')
-                
-                window.location.href = "/"
 
             } catch (error) {
                 console.error('Erreur lors de la suppression de la session:', error.message);
@@ -115,16 +159,26 @@ export default function Profil({ user }){
         <>
         <div className={styles.Profil}>
             <span className={styles.Title}>Vos informations personnelles</span>
-            <div className="mTop30">
-                <div className="flex gap20 toColumn wm100">
-                    <input disabled type="text" name="nom" onChange={handleChange} value={userData?.nom} className="input-text w50 wm100" placeholder="Nom" />
-                    <input disabled type="text" name="prenom" onChange={handleChange} value={userData?.prenom} className="input-text w50 wm100" placeholder="Prénom" />
-                </div>
-                <div className="flex gap20 mTop20 toColumn">
-                    <input disabled type="mail" name="mail" onChange={handleChange} value={userData?.mail} className="input-mail w50 wm100" placeholder="Adresse e-mail" />
-                    <input disabled type="text" name="telephone" onChange={handleChange} value={userData?.telephone} className="input-text w50 wm100" placeholder="Numéro de téléphone" />
-                </div>
-            </div>
+            {user.type != "Administrateur" && user.type != "DR" ? (
+                <>
+                    <div className="mTop30">
+                        <div className="flex gap20 toColumn wm100">
+                            <input disabled type="text" name="nom" onChange={handleChange} value={userData?.nom} className="input-text w50 wm100" placeholder="Nom" />
+                            <input disabled type="text" name="prenom" onChange={handleChange} value={userData?.prenom} className="input-text w50 wm100" placeholder="Prénom" />
+                        </div>
+                        <div className="flex gap20 mTop20 toColumn">
+                            <input disabled type="mail" name="mail" onChange={handleChange} value={userData?.mail} className="input-mail w50 wm100" placeholder="Adresse e-mail" />
+                            <input disabled type="text" name="telephone" onChange={handleChange} value={userData?.telephone} className="input-text w50 wm100" placeholder="Numéro de téléphone" />
+                        </div>
+                    </div>                
+                </>
+            ) : (
+                <div className="mTop30">
+                    <div className="flex gap20 mTop20 toColumn">
+                        <input disabled type="mail" name="mail" onChange={handleChange} value={userData?.email} className="input-mail w50 wm100" placeholder="Adresse e-mail" />
+                    </div>
+                </div>  
+            )}
             <span className={styles.Subtitle}>Vous souhaitez modifier votre mot de passe ?</span>
             <div className="mTop30">
                 <div className="flex gap20 toColumn">
